Sort the mime types before comparing
For those tests that use start we don't need to assert the actual order
of mime types that are returned.

This happen because this order is more about the order the mime type was
registered than the order that it is expected to it resolve.

We need to sort because we remove the json mime type in
json_params_parsing_test and add it to the end of the mime types set so
if that file runs before those tests we will have a failing test.

[Rafael Mendonça França + Lucas Hosseini]